requirements.txt
gitpython==3.1.41
boto3>=1.28.0
